requirements.txt
langchain>=0.1.17
langchain-community>=0.0.21
langchain-core>=0.1.50
langchain-google-genai>=0.0.11
google-generativeai>=0.3.2
streamlit>=1.32.0
python-dotenv>=1.0.1
ollama>=0.1.27
